refactor(events): relocate validation/consensus KM reactions to knowledge domain home (P4a Batch E2b)
Move the live validation/consensus/provenance Knowledge Mound event
reactions out of the infrastructure event bus into the knowledge domain
home (aragora/knowledge/event_subscribers.py), self-registering via the
E1 registry API, per docs/architecture/P4A_EVENTS_QUEUE_INVERSION.md
(section 10, row E2 / E2b).

- Relocate _handle_provenance_to_mound, _handle_mound_to_provenance,
  _handle_consensus_to_mound, _handle_km_validation_feedback into
  KnowledgeEventSubscriber; extend the handler-name frozenset to 15 and
  register the 4 reactions in KnowledgeEventSubscriber.register().
- De-mixin CrossSubscriberManager: drop ValidationHandlersMixin and its
  4 builtin register blocks; the bare manager stays infrastructure-only.
- Delete aragora/events/cross_subscribers/handlers/validation.py and the
  dead-duplicate aragora/events/subscribers/mound_handlers.py (never
  composed into CrossSubscriberManager; deletion avoids a double-fire).
  No re-export shim (Relocate-UP no-shim exemption).
- Repoint all consumers: test_consensus_ingestion, test_e2e_debate_km_flow,
  test_km_bidirectional (provenance), and the registry parity test.

get_cross_subscriber_manager() import path unchanged.
