+++ aragora/events/cross_subscribers/handlers/__init__.py
@@ -7,12 +7,10 @@ organized by subsystem category.
 
 from .basic import BasicHandlersMixin
 from .culture import CultureHandlersMixin
-from .validation import ValidationHandlersMixin
 from .strategic import StrategicHandlersMixin
 
 __all__ = [
     "BasicHandlersMixin",
     "CultureHandlersMixin",
-    "ValidationHandlersMixin",
     "StrategicHandlersMixin",
 ]
