+++ tests/integration/test_km_bidirectional.py
@@ -393,10 +393,10 @@ class TestProvenanceBidirectional:
 
     def test_provenance_to_mound_stores_verified(self):
         """Test that verified provenance chains are stored."""
-        from aragora.events.cross_subscribers import CrossSubscriberManager
+        from aragora.knowledge.event_subscribers import KnowledgeEventSubscriber
         from aragora.events.types import StreamEvent, StreamEventType
 
-        manager = CrossSubscriberManager()
+        subscriber = KnowledgeEventSubscriber()
 
         event = StreamEvent(
             type=StreamEventType.CONSENSUS,
@@ -416,14 +416,14 @@ class TestProvenanceBidirectional:
         )
 
         # Should not raise
-        manager._handle_provenance_to_mound(event)
+        subscriber._handle_provenance_to_mound(event)
 
     def test_provenance_to_mound_ignores_no_consensus(self):
         """Test that provenance is not stored without consensus."""
-        from aragora.events.cross_subscribers import CrossSubscriberManager
+        from aragora.knowledge.event_subscribers import KnowledgeEventSubscriber
         from aragora.events.types import StreamEvent, StreamEventType
 
-        manager = CrossSubscriberManager()
+        subscriber = KnowledgeEventSubscriber()
 
         event = StreamEvent(
             type=StreamEventType.CONSENSUS,
@@ -434,14 +434,14 @@ class TestProvenanceBidirectional:
         )
 
         # Should not raise
-        manager._handle_provenance_to_mound(event)
+        subscriber._handle_provenance_to_mound(event)
 
     def test_mound_to_provenance_queries_history(self):
         """Test that claim verification queries KM for history."""
-        from aragora.events.cross_subscribers import CrossSubscriberManager
+        from aragora.knowledge.event_subscribers import KnowledgeEventSubscriber
         from aragora.events.types import StreamEvent, StreamEventType
 
-        manager = CrossSubscriberManager()
+        subscriber = KnowledgeEventSubscriber()
 
         event = StreamEvent(
             type=StreamEventType.CLAIM_VERIFICATION_RESULT,
@@ -453,7 +453,7 @@ class TestProvenanceBidirectional:
         )
 
         # Should not raise
-        manager._handle_mound_to_provenance(event)
+        subscriber._handle_mound_to_provenance(event)
 
 
 class TestRankingAdapter:
